pyproject.toml
[tool.poetry]
name = "cantina"
version = "0.1.0"
description = "Sistema de gestão para cantinas"
authors = ["Timóteo Ramos <timoteoramos@ymail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
django-environ = "^0.11.2"
django-localflavor = "^4.0"
django-model-utils = "^5.0.0"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "^8.0.0" }
uvicorn = "^0.34.0"
psycopg = { extras = ["binary"], version = "^3.2.3" }
whitenoise = { extras = ["brotli"], version = "^6.8.2" }

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.1"
pylint-django = "^2.6.1"

[tool.poetry.scripts]
dev = "cantina.scripts:devserver"
django = "cantina.manage:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
